Fixed spcode
Fixed spcode in water_potential.csv. The burn_trials.csv might show
rewritten but no input is changed. I was getting lots of NA after
running the hobos data in R and I was trying to figure it out. I thought
it might be result of chnaging the data type in same column!! Sometimes
it happens in my laptop!! particularly date. As a result I copied the
burn_trails.csv and made a fresh version!